Modified ErrorCPD because of php strict warning.
ErrorCPD had a getDescription method with different parameters than it's
parent class PluginsAbstract, which caused a warning when strict mode
was enabled.

Because of this, the old getDescription method was moved to
_CPDgetDescription and a new getDescription method that is defined
according to PluginsAbstract->getDescription was added. When called,
this method throws an exception, stating that the operation is not
supported.