Fixed a possible bug when handling files in very different locations.
The _getTreeListHtml function used directory strings without trailing
slashes. This caused $curDir to become the empty string, which in turn
caused strpos($dir, $curDir) to return false.

This problem was solved by adding trailing slashes to directories.
Now, when we hit the root directory, $curDir is DIRECTORY_SEPARATOR and
everything is fine.