Removed debug output and thus the dependencies on Log and PHP_Timer.
The inclusion of Log and PHP_Timer had 3 reasons.
1. Debug output. This did not contain any useful information and isn't
neccessary for a script that has a linear way of execution like this
one. If neccessary, simple echo statements added on demand will do the
trick for devs.

2. Performance output. As the ouput did not show fractions of a second,
this was not so useful anyways. XDebug is a way better tool to look at
performance issues.

3. Error output. These are replaced by php's own error_log function,
which works exactly as well.

As all of this reasons are not really good ones, the logging was
removed.