Added alternating colors for tables. Closes #20.
While on it, I also grabbed the alternating colors css for ols that was
used in the reviews and put both in a new css file, global.css. (It's
the same code.)
I also fixed a naming inconsistency. On the main page, rows had the
classes 'file' and 'oddrow', in reviews they were called 'odd' and
'white'.

Now they're 'odd' and 'even' in both places.