Fixing broken formatting with non-php files without Text_Highlighter installed. Closes #40.
This fixes two problems with the highlighting. The .* matched both the
whole line and the newline, which left us with an empty line after each
line of the source file.
The other problem was that newlines weren't shown in browsers because
they weren't replaced with &nbsp;.

Thanks to galkin for spotting this one.